+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HCI: fix report of extended advertisements with data len > 31
 - HCI: fix CTKD in Initiator role over BR/EDR SC when using dedicated bonding
 - SM: fix value in SM_EVENT_NUMERIC_COMPARISON_REQUEST
+- SM: ignore Security Request during re-encryption
 - btstack_stdin_embedded: use timer to poll RTT input, fix for tickless RTOS
 - gatt_client: return ERROR_CODE_UNKNOWN_CONNECTION_IDENTIFIER for invalid connection handle
 - A2DP: emit A2DP_SUBEVENT_STREAM_ESTABLISHED with if connection fails, e.g. because of Page Timeout